pyproject.toml
[project]
name = "pypi-simple-server"
version = "2.3.1"
description = 'Simple PEP503 index for python packages'
authors = [
  {name = "Sebastian Koslowski", email = "s.koslowski@procitec.de"},
]
license = "MIT"
requires-python = ">=3.14"
dependencies = [
  "humanize>=4.12.2",
  "jinja2>=3.1.5",
  "msgspec>=0.19.0",
  "packaging>=24.2",
  "pyjwt>=2.10.1",
  "python-multipart>=0.0.22",
  "starlette>=0.45.3",
  "uvicorn[standard]>=0.15.0",
  "watchfiles>=1.0.4",
]

[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"

[tool.ruff] # https://beta.ruff.rs/docs/rules
lint.extend-select = [
  "I", # isort
  "UP", # pyupgrade
  "PT", # flake8-pytest-style
]
line-length = 110

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pypi-simple>=1.6.1",
    "pytest-cov>=6.0.0",
    "pytest>=8.3.4",
    "ruff>=0.9.6",
    "types-pyjwt>=1.7.1",
]

[tool.pytest]
addopts = [
    "--cov=pypi_simple_server",
    "--import-mode=importlib",
]
testpaths = [
    "tests",
]
